pyproject.toml
[project]
name = "stayflex"
version = "0.1.0"
description = "StayFlex - A voucher marketplace for hotel experience packages"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "drf-spectacular>=0.27,<0.28",
    "psycopg[binary]>=3.1,<4.0",
    "python-dotenv>=1.0,<2.0",
    "requests>=2.31,<3.0",
    "phonenumbers>=8.13,<9.0",
    "django-cors-headers>=4.3,<5.0",
    "whitenoise>=6.6,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.8,<5.0",
    "black>=24.0,<25.0",
    "ruff>=0.3,<0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
